good_Tokenとgood_ASTの定義を各.hファイルへ分離
